Fix preview for files in projects
This patch aims to fix #13.
When opening a project, each file-editor receives a pointer to the editor
instance of the root file. After saving a non-root file of a project, Gummi will
compile the pdf as if the root file was changed.

Note: This doesn't provide a live-preview for non-root files, but at least it
shows the compiled pdf, without having to switch to the root-file editor.

I did test it a bit (on Linux only though), however since I didn't know the
codebase at all before doing the change, I don't know whether I missed
something.